Chore: ESLINT v10 adjustments
